requirements.txt
Babel==1.3
BeautifulSoup==3.2.1
CherryPy==3.2.2
Django==1.6
Jinja2==2.7.2
Mako==0.9.1
Markdown==2.4
MarkupSafe==0.18
MySQL-python==1.2.3
PAM==0.4.2
Pillow==2.3.0
PyAudio==0.2.7
PyBluez==0.18
PyChart==1.39
PyOpenGL==3.0.2
PyRoom==0.4.1
PyYAML==3.10
Pygments==1.6
Routes==2.0
Twisted-Core==13.2.0
Twisted-Web==13.2.0
UniConvertor==1.1.4
WebOb==1.3.1
Werkzeug==0.9.4
adium-theme-ubuntu==0.3.4
## FIXME: could not find svn URL in dependency_links for this package:
apsw==3.8.2-r1
apt-xapian-index==0.45
argparse==1.2.1
ccsm==0.9.11.3
chardet==2.0.1
classicmenu-indicator==0.07
colorama==0.2.5
command-not-found==0.3
compizconfig-python==0.9.11.3
cssselect==0.9.1
cssutils==0.9.10
debtagshw==0.1
decorator==3.4.0
defer==1.0.6
dirspec==13.10
dnspython==1.11.1
docutils==0.11
duplicity==0.6.23
feedparser==5.1.3
gevent==1.0
greenlet==0.4.2
gyp==0.1
html5lib==0.999
httplib2==0.8
ipython==1.2.1
lockfile==0.8
lxml==3.3.3
matplotlib==1.3.1
mechanize==0.2.5
mock==1.0.1
mysql-connector-python==1.1.6
mysql-utilities==1.3.5
netifaces==0.8
numpy==1.8.2
oauthlib==0.6.1
oneconf==0.3.7.14.04.1
paramiko==1.10.1
passlib==1.5.3
pexpect==3.1
piston-mini-client==0.7.5
psutil==1.2.1
psycopg2==2.4.5
pyOpenSSL==0.13
pyPdf==1.13
pycrypto==2.6.1
pycups==1.9.66
pycurl==7.19.3
pydot==1.0.28
pygame==1.9.1release
pygobject==3.12.0
pyinotify==0.9.4
pyogg==1.3
pyparsing==2.0.1
pyserial==2.6
pysmbc==1.0.14.1
pysqlite==2.6.3
python-apt==0.9.3.5ubuntu1
python-dateutil==1.5
python-debian==0.1.21-nmu2ubuntu2
python-ldap==2.4.10
python-openid==2.2.5
pytz==2012c
pyvorbis==1.5
pyxdg==0.25
reportlab==3.0
repoze.lru==0.6
requests==2.2.1
roman==2.0.0
sessioninstaller==0.0.0
simplegeneric==0.8.1
simplejson==3.3.1
six==1.5.2
software-center-aptd-plugins==0.0.0
ssh-import-id==3.21
suds==0.4.1
system-service==0.1.6
uTidylib==0.2
unittest2==0.5.1
unity-lens-photos==1.0
urllib3==1.7.1
vatnumber==1.1
virtualenv==13.1.2
vobject==0.8.1c
wheel==0.24.0
wsgiref==0.1.2
wxPython==2.8.12.1
wxPython-common==2.8.12.1
xdiagnose==3.6.3build2
xdot==0.5
xlwt==0.7.5
zope.interface==4.0.5
